fix overflow

+++ MeansToAnEnd.cs
@@ -1,5 +1,6 @@
 using System.Buffers;
 using System.Buffers.Binary;
+using System.Numerics;
 using protohackers.Transport;
 
 namespace protohackers;
@@ -76,7 +77,7 @@ public class MeansToAnEnd : TcpServerBase
                 var filtered = prices.GetViewBetween(
                     new TimestampedPrice(low, 0),
                     new TimestampedPrice(high, 0));
-                long sum = 0;
+                BigInteger sum = 0;
                 foreach (var p in filtered)
                 {
                     sum += p.Price;
